requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.9.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.3.1
protobuf==6.33.0
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
streamlit-drawable-canvas==0.9.3
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wheel==0.45.1
